etl/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-core==2.0.0
google-cloud-datastore==2.1.6
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
libcst==0.3.21
mypy-extensions==0.4.3
proto-plus==1.19.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
